Posts fetch, markdown is broken
Posts do fetch, but as I was writing the fetch function I thought that it would be quite a bother to have postcontent.js be this massive file that handles all the markdown (including any future markdown I might want). Then I thought, since I have this ability to write custom elements fully now I might as well take advantage of it by having custom elements actually become the markdown. My plan for the future is to implement markdown blocks which handle raw text inside of them. This allows me to use different markdowns, and if implemented in the way I'm thinking will allow me to switch markdowns. This would make it easier to implement syntax highlighting or a consistant display for post titles. I feel the content of the blog will get a major overhaul here.